+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-06-03 21:33:14
-// Commit: perf(perf): optimize memory consumption in perf list viewport
+// Updated at: 2026-06-04 10:37:51
+// Commit: fix(perf): prevent double tap duplicate action in perf
 
 class perfComponentState {
-  final String commitRef = '1731';
-  final String lastMessage = 'perf(perf): optimize memory consumption in perf list viewport';
-  final DateTime timestamp = DateTime.parse('2026-06-03T21:33:14+05:00');
+  final String commitRef = '1743';
+  final String lastMessage = 'fix(perf): prevent double tap duplicate action in perf';
+  final DateTime timestamp = DateTime.parse('2026-06-04T10:37:51+05:00');
 }