+++ CHANGELOG.md
@@ -1744,3 +1744,4 @@
 - [2026-06-03 23:00] test(auth): add unit tests for auth business logic
 - [2026-06-03 23:26] test(i18n): implement widget integration tests for i18n
 - [2026-06-04 10:03] test(theme): implement widget integration tests for theme
+- [2026-06-04 10:37] fix(perf): prevent double tap duplicate action in perf